+++ app/src/client/components/RemoveLineFromSubscriptionButton.tsx
@@ -8,6 +8,7 @@ import {
   REMOVE_SUBSCRIPTION_DRAFT_LINE,
   COMMIT_SUBSCRIPTION_DRAFT,
 } from '../handlers';
+import { calculateCurrentPrice, calculateDiscountRate } from '../utils/subscription-box';
 
 interface Props {
   contractId: string;
@@ -19,12 +20,6 @@ interface Props {
   setToastError: (error: boolean) => void;
   refetch: () => void;
 }
-
-const calculateCurrentPrice = (discountRate: number, currentPrice: string) => {
-  const price = parseFloat(currentPrice);
-  const amount = (price - price * discountRate).toFixed(2);
-  return String(amount);
-};
 
 const RemoveLineFromSubscriptionButton = ({
   contractId,
@@ -45,18 +40,7 @@ const RemoveLineFromSubscriptionButton = ({
     if (lineId !== line.node.id) totalQuantity += line.node.quantity;
   });
   // move this to constant later
-  let discountRate: number = 0;
-  if (isSubscriptionBox) {
-    if (totalQuantity >= 5) {
-      discountRate = 0.2;
-    } else if (totalQuantity >= 4) {
-      discountRate = 0.15;
-    } else if (totalQuantity >= 3) {
-      discountRate = 0.1;
-    } else {
-      discountRate = 0;
-    }
-  }
+  const discountRate = calculateDiscountRate(totalQuantity, isSubscriptionBox);
   // get updated pricing per line
   const linesWithUpdatedPrices: any[] = [];
   lineItems.forEach((line: any) => {
